Remove wrapper functions from handlers table
Before, the handlers table contained wrapper functions. For example: handlers.chat was an unnamed function that took two arguments, msg and console, and called handleChat() passing it those two arguments. This is totally unnecesary, and ends up causing a problem with returning values. handlePing() was returning the pong message and handle.message() was expecting that return value to send on to the IRC thread for sending. However because of the wrapper function, the pong message was returned to the wrapper function which did nothing with it so handle.message() never got it and never sent it on to the IRC thread. Because of this oversight, the bot couldn't pong and would get disconnected from channels due to timeout.